Address https://github.com/ovro-lwa/lwa-issues/issues/486#issuecomment-1927325521
The X-engine baseline selection happens on a per-polarization level - i.e.
it is possible to select baselines of arbitrary polarization, e.g. all XX pols.
The fast correlator output has two packet formats, one allows arbitrary selection
while the other is COR-standard and assumes that each packet contains data from
all 4 polarization-products of a station pair.
If using this output format (which the LWA-OVRO correlator typically does) the baselines
selected are assumed to be in sets of 4 where each 4 consecutive baselines are the 4 pol-pairs
of a single station pair. In this case, the 4 baselines selected must match the polarization
order expeted by the COR format.